[refactor] Add `UpdateTaskHandler` abstraction
- package base command handler into `command.handler.base`
