+++ src/main/java/duke/command/handler/CommandTagHandler.java
@@ -4,13 +4,13 @@ import java.util.regex.MatchResult;
 import java.util.regex.Pattern;
 
 import duke.command.CommandException;
+import duke.command.handler.base.CommandUpdateTaskHandler;
 import duke.command.response.CommandResponse;
 import duke.command.response.UpdateTaskResponse;
-import duke.command.response.UpdateTaskResponse.UpdateType;
 import duke.data.TaskList;
 import duke.data.tasks.Task;
 
-public class CommandTagHandler extends CommandHandler {
+public class CommandTagHandler extends CommandUpdateTaskHandler {
 
     protected static final String INVALID_FORMAT_MESSAGE = String.join("\n",
         "Invalid `tag` command format!",
@@ -30,24 +30,17 @@ public class CommandTagHandler extends CommandHandler {
     }
 
     @Override
-    public CommandResponse run(TaskList taskList) throws CommandException {
+    protected String getSelectedTaskIdStr() {
         MatchResult regexMatchResult = commandRegexMatcher.toMatchResult();
+        return regexMatchResult.group(2);
+    }
 
+    @Override
+    protected CommandResponse updateTask(Task task, int taskIdx, TaskList taskList) {
+        MatchResult regexMatchResult = commandRegexMatcher.toMatchResult();
         String tagName = regexMatchResult.group(1);
-        String taskIdxStr = regexMatchResult.group(2);
 
-        try {
-            int taskIdx = Integer.parseInt(taskIdxStr);
-            if (taskIdx <= 0 || taskIdx > taskList.size()) {
-                throw new CommandException("Invalid task selected!");
-            }
-            Task task = taskList.getTask(taskIdx - 1);
-            task.addTag(tagName);
-            return new UpdateTaskResponse(task, UpdateType.TAG);
-        } catch (NumberFormatException error) {
-            throw new CommandException(String.join("\n",
-                "Task number should be a number!",
-                "Got: %s", taskIdxStr));
-        }
+        task.addTag(tagName);
+        return new UpdateTaskResponse(task, UpdateTaskResponse.UpdateType.TAG);
     }
 }
